refactor locales
