+++ samanthas_telegram_bot/conversation/callbacks/registration/teacher_under_18.py
@@ -10,8 +10,8 @@ from samanthas_telegram_bot.conversation.auxil.enums import (
     ConversationStateTeacherUnder18,
 )
 from samanthas_telegram_bot.conversation.auxil.helpers import answer_callback_query_and_get_data
-from samanthas_telegram_bot.data_structures.constants import Locale
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
+from samanthas_telegram_bot.data_structures.literal_types import Locale
 
 
 async def ask_readiness_to_host_speaking_club(
