+++ samanthas_telegram_bot/data_structures/literal_types.py
@@ -1,0 +1,4 @@
+from typing import Literal
+
+CommunicationModeInClass = Literal["ru", "ua", "ru_ua", "l2_only"]
+Locale = Literal["ua", "en", "ru"]
